refactor(common/lmlayer): move placeholder word breaker to @keymanapp/lexical-model-word-breakers
