+++ common/lexical-model-word-breakers/test/test-placeholder-word-breaker.ts
@@ -1,13 +1,13 @@
 /// <reference types="@keymanapp/lexical-model-types" />
 
 import {assert} from 'chai';
-import {placeholder as breakWords} from '../'
+import {placeholder as breakWords} from '../src';
 
 
 describe('The placeholder word breaker', function () {
   it('should break simple English sentences', function () {
     let breaks = breakWords('Look! -- The quick brown fox jumps... over the lazy dog!');
     let words = breaks.map(span => span.text);
-    assert.deepEqual(words, ['Look', 'The', 'quick', 'brown', 'fox', 'jumps', 'over', 'the', 'lazy', 'dog']);
+    assert.deepEqual(words, ['Look!', '--', 'The', 'quick', 'brown', 'fox', 'jumps...', 'over', 'the', 'lazy', 'dog!']);
   });
 });
